fix some x64 build breaks

+++ src/priv/gfx/gpu/d3d10/d3d10RenderTargetMgr.h
@@ -64,7 +64,7 @@ namespace GN { namespace gfx
         ///
         /// get render target count
         ///
-        UInt32 getRenderTargetCount() const { return mNumColors; }
+        UInt32 getRenderTargetCount() const { return (UInt32)mNumColors; }
 
         ///
         /// get render target view
